refactor: ChessGame에서 MoveDto 사용하도록 수정
